requirements.txt
# Core dependencies
pandas>=2.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=3.3.5

# Data processing
pyyaml>=6.0
joblib>=1.3.0
pytz>=2023.3
pyarrow>=19.0.0

# Web and API
requests>=2.25.0
streamlit==1.28.0
altair==4.2.2
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
watchdog>=6.0.0
protobuf>=4.25.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Production dependencies
gunicorn>=21.0.0
uvicorn>=0.23.0
fastapi>=0.100.0
pydantic>=2.0.0

# Monitoring and logging
structlog>=23.1.0
python-json-logger>=2.0.0
prometheus-client>=0.17.0
